Revert "moved WorkflowProcess to de.uka.ipd.sdq.workflow.launchconfig.core"
This reverts commit e4129378f5016d94cf0fa76bb9bf123d0ef0ad29.
