+++ bundles/de.uka.ipd.sdq.workflow.helloworld/META-INF/MANIFEST.MF
@@ -8,5 +8,4 @@ Bundle-RequiredExecutionEnvironment: JavaSE-17
 Require-Bundle: de.uka.ipd.sdq.workflow;bundle-version="2.0.0",
  de.uka.ipd.sdq.workflow.launchconfig;bundle-version="2.0.0",
  org.eclipse.jface;bundle-version="3.10.1",
- de.uka.ipd.sdq.workflow.workbench;bundle-version="2.1.0",
- de.uka.ipd.sdq.workflow.launchconfig.core;bundle-version="5.2.0"
+ de.uka.ipd.sdq.workflow.workbench;bundle-version="2.1.0"
