+++ bundles/de.uka.ipd.sdq.workflow.workbench/META-INF/MANIFEST.MF
@@ -9,8 +9,7 @@ Require-Bundle: org.eclipse.debug.core;bundle-version="3.5.0";visibility:=reexpo
  de.uka.ipd.sdq.workflow.ui;bundle-version="1.0.0";visibility:=reexport,
  org.eclipse.ui;bundle-version="3.5.1";visibility:=reexport,
  org.eclipse.ui.console;bundle-version="3.4.0",
- de.uka.ipd.sdq.workflow;bundle-version="1.1.0";visibility:=reexport,
- de.uka.ipd.sdq.workflow.launchconfig.core;bundle-version="5.2.0"
+ de.uka.ipd.sdq.workflow;bundle-version="1.1.0";visibility:=reexport
 Export-Package: de.uka.ipd.sdq.workflow.workbench
 Bundle-Vendor: sdq.ipd.uka.de
 Bundle-RequiredExecutionEnvironment: JavaSE-17
